Enhanced Appearance
Scrolling Bitmap - A spaceship remains in the center of the screen as the stars behind it scroll. Demonstrates scrolling an image that wraps around at the edges to allow infinite scrolling in any direction.

Parallax Scroll - A spaceship remains in the center of the screen. The planets behind the ship and the stars scroll at different speeds to provide the illusion of depth. Demonstrates parallax scrolling to create the illusion of depth.

Shadow - An airplane flies over the ocean with its shadow visible on the surface of the water. Demonstrates using alpha transparency to create a shadow.

Reflection - An airplane flies over the ocean with its reflection visible on the surface of the water. Demonstrates using alpha transparency to create a reflection.

Message Dialog - A message dialog is displayed on the screen with a test message and an OK button. Clicking the OK button or pressing the Enter key closes the dialog. Demonstrates how to display the Message Dialog.

Input Dialog - An input dialog is displayed on the screen with a message and a text input region. Entering the word ‘next’ displays a second dialog. Demonstrates how to use the Input Dialog.

Dashboard Demo - A dashboard of gauges is displayed. Demonstrates how to display each dashboard element.